[Chore] Separate out `vllm.utils.collections`
