+++ tests/conftest.py
@@ -57,7 +57,8 @@ from vllm.multimodal.utils import fetch_image
 from vllm.outputs import RequestOutput
 from vllm.sampling_params import BeamSearchParams
 from vllm.transformers_utils.utils import maybe_model_redirect
-from vllm.utils import is_list_of, set_default_torch_num_threads
+from vllm.utils import set_default_torch_num_threads
+from vllm.utils.collections import is_list_of
 
 logger = init_logger(__name__)
