+++ vllm/entrypoints/openai/serving_embedding.py
@@ -39,8 +39,8 @@ from vllm.outputs import (
     RequestOutput,
 )
 from vllm.pooling_params import PoolingParams
-from vllm.utils import chunk_list
 from vllm.utils.asyncio import merge_async_iterators
+from vllm.utils.collections import chunk_list
 
 logger = init_logger(__name__)
